reply-bot/posting/requirements.txt
openai
python-dotenv
tweepy